coluna total criada

+++ src/database/migrations/20230329171329_products.js
@@ -2,7 +2,7 @@ exports.up = function(knex) {
   return knex.schema.createTable('products', (table) => {
     table.increments('id');
     table.string('title').notNullable;
-    table.float('price').notNullable;
+    table.decimal('price').notNullable;
   })
 };
