+++ src/database/migrations/20230413110217_user-products-column-total.js
@@ -1,6 +1,6 @@
 exports.up = function(knex) {
   return knex.schema.alterTable('user_products', (table) => {
-    table.integer('total').notNullable;
+    table.decimal('total').notNullable;
   })
 };
